requirements.txt
python-telegram-bot[job-queue]==20.7
supabase==2.3.4
pytz
python-dotenv
